% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agi.R
\name{mod_agi_server}
\alias{mod_agi_server}
\title{Function to plot AGI in aggregate or per cap}
\usage{
mod_agi_server(input, output, session, dataset, var)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate.}

\item{dataset}{data frame (non-reactive) with variables necessary}

\item{var}{reactive list of with variable type}
}
\description{
Function to plot AGI in aggregate or per cap
}
