% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agi.R
\name{mod_agi_ui}
\alias{mod_agi_ui}
\title{Function to plot AGI in aggregate or per cap}
\usage{
mod_agi_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
A shiny Module.
}
