% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_varselect.R
\name{mod_varselect_server}
\alias{mod_varselect_server}
\title{varselect Server Function}
\usage{
mod_varselect_server(input, output, session)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate.}
}
\value{
list with following components
\describe{
  \item{type}{reactive character indicating type selection}
}
}
\description{
varselect Server Function
}
