% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_varselect.R
\name{mod_varselect_ui}
\alias{mod_varselect_ui}
\title{varselect UI Function}
\usage{
mod_varselect_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
A shiny Module.
}
